[feature #163905397] refactored index endpoint
