+++ app/api/v2/__init__.py
@@ -1,0 +1,4 @@
+# """Initiates blueprints for the API"""
+# from flask import Blueprint
+
+# version_1 = Blueprint('version_one', __name__, url_prefix='/api/v1')
