+++ app/api/v2/models/base_model.py
@@ -1,7 +1,15 @@
 from datetime import datetime
-from ....database_config import init_db
+from ..utils.utils import generate_id
 
-class BaseModel(object):
+class Model(object):
+  """
+    Base model class
+  """
+
+  from datetime import datetime
+from ..utils.utils import generate_id
+
+class Model(object):
   """
     Base model class
   """
@@ -20,36 +28,29 @@ class BaseModel(object):
     self.collection.append(data)
     return data
 
-  def check_if_it_exists(self, table_name, field_name, value):
+  def check_if_it_exists(self, key, value):
     """
       method to check if an object exist using key, value pair
     """
-    con = init_db()
-    cur = con.cursor()
-    query = "SELECT * FROM {} WHERE {}='{}';".format(table_name, field_name, value)
-    cur.execute(query)
-    resp = cur.fetchall()
-    if resp:
-      return True
-    else:
-      return False
+    items = [item for item in self.collection if item[key] == value]
+    return len(items) > 0
 
-  # def find(self, key, value):
-  #   """
-  #     method to find object item using key, value pair
-  #   """
-  #   items = [item for item in self.collection if item[key] == value]
-  #   return items[0]
+  def find(self, key, value):
+    """
+      method to find object item using key, value pair
+    """
+    items = [item for item in self.collection if item[key] == value]
+    return items[0]
 
-  # def fetch_all(self):
-  #   """
-  #     method to fetch all items objects
-  #   """
-  #   return self.collection
+  def fetch_all(self):
+    """
+      method to fetch all items objects
+    """
+    return self.collection
 
-  # def delete(self, id):
-  #   """
-  #     method to delete item object
-  #   """
-  #   item = self.find('id', id)
-  #   self.collection.remove(item)
+  def delete(self, id):
+    """
+      method to delete item object
+    """
+    item = self.find('id', id)
+    self.collection.remove(item)
