+++ app/api/v2/models/political_party_model.py
