+++ app/api/v2/schemas/political_party_schema.py
