+++ app/api/v2/schemas/user_schema.py
@@ -1,6 +1,6 @@
 from marshmallow import Schema, fields, post_dump
-from ...v1.utils.validators import email, password, required
-# from ..models.user_model import User
+from ..utils.validators import email, password, required
+from ..models.user_model import User
 
 class UserSchema(Schema):
   """
