+++ app/api/v2/views/office_view.py
